Update README.md
A piece of moving art work inspired by triangles. Each side of the landscape will be cover in a landscape from nature. Each one with a moving element on it. One side will have an ocean environment with moving waves, another with a forest environment with swaying trees, and another with a mountain environment with moving clouds.